refactor(store): rename mime reducer to `shapeReducer` because it only handles shapes
