+++ browser/store/reducers/shapeReducer.js
